+++ CHANGELOG.md
@@ -6,8 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
-- WIP
+## 2026 Update [1.0.1](https://github.com/TangoMan75/calculatrice-salaire-net/releases/tag/1.0.1)
+### Added
+- AGENTS.md
+- Reference wages for years 2026 and 2025 in `src/data/reference_wages.json`.
+### Changed
+- Copyright year in `LICENSE` to 2026 and in `src/components/Card.vue` footer.
+- Unit tests for `footer.spec.js` to reflect new copyright and version.
 
-## [0.1.0](https://github.com/TangoMan75/calculatrice-salaire-net/releases/tag/0.1.0)
+## Initial Release [1.0.0](https://github.com/TangoMan75/calculatrice-salaire-net/releases/tag/1.0.0)